spawn: preserve full 32-bit exit codes on Windows
Windows exit codes are a full 32-bit DWORD and NTSTATUS crash codes
(0xC0000005 access violation, 0xC0000374 heap corruption, etc.) live in
the high bits. Storing the code as u8 truncated every crash code to a
meaningless small integer (0xC0000005 became 5), indistinguishable from
a normal process.exit(5). Node reports the full value.

Widen Exited::code from u8 to u32 and stop truncating in the libuv exit
callback. POSIX WEXITSTATUS is still 8 bits so behavior there is
unchanged.

+++ src/runtime/cli/pm_version_command.rs
@@ -823,8 +823,8 @@ impl PmVersionCommand {
             }
             Ok(result) => {
                 if !result.is_ok() {
-                    let exit_code: i32 = match &result.status {
-                        ProcStatus::Exited(e) => i32::from(e.code),
+                    let exit_code: i64 = match &result.status {
+                        ProcStatus::Exited(e) => i64::from(e.code),
                         _ => -1,
                     };
                     Output::err_generic("Git add failed with exit code {}", (exit_code,));
